Change yeoman.out to .tmp
The yeoman configurable paths are intended to be used for paths that change. For
example, you might have a lab folder for building and syncing to lab, and a dist
folder for building and syncing to prod. In reality, there tend to be a couple
of places where the configurable path names are hard coded, maybe in spec files,
maybe in settings files (like .bowerrc). The obj might be a relic from the past,
and maybe we should PR it out of Yeoman in general.

Anyways, there's not much reason to configure a staging directory for compiled
files, so switching it to the conventionally hardcoded .tmp.